+++ public_html/index.php
@@ -43,7 +43,7 @@ if ($access & WEATHER_ACCESS_SEARCH) {
     // Otherwise, get the user's profile location (non-anonymous only).
     if (isset($_REQUEST['loc']) && !empty($_REQUEST['loc'])) {
         // Get the requested location if submitted
-        $loc = WEATHER_stripslashes($_REQUEST['loc']);
+        $loc = $_REQUEST['loc'];
     }
 }
 
@@ -72,34 +72,42 @@ $Session = SESS_getVar('glWeather');
 if ($Session === 0) $Session = array();
 
 if (!empty($loc)) {
-    $key = array_search($loc, $Session);
-    if ($key !== false) {
-        unset($Session[$key]);
+    $found = false;
+    $l_loc = strtolower($loc);
+    foreach ($Session as $key=>$value) {
+        if (strtolower($value) == $l_loc) {
+            $found = true;
+            break;
+        }
     }
-    $Session[] = $loc;
+    if (!$found) {
+        array_unshift($Session, $loc);
+    }
 }
-SESS_setVar('glWeather', $Session);
 
 $T = new Template(WEATHER_PI_PATH . '/templates');
 $T->set_file('index', 'index.thtml');
 $T->set_block('index', 'EmbedBlock', 'eBlk');
 $msg = '';
 foreach ($Session as $key=>$loc) {
-    $weather = WEATHER_getWeather($loc);
+    $weather = Weather\api::getWeather($loc);
     if (!is_array($weather)) {
         unset($Session[$key]);    // how'd bad weather get here?
-        $msg = $weather;
+        //$msg = $weather;
         continue;
     }
     $embed = WEATHER_embed($weather, false);
     $T->set_var(array(
         'embed' => $embed,
         'divid' => $key,
-        'linkback' => Weather::linkback(),
+        'linkback' => Weather\api::linkback(),
         'iconset' => $_CONF_WEATHER['_iconset'],
     ) );
     $T->parse('eBlk', 'EmbedBlock', true);
 }
+// Save the session var. Should have only valid locations now.
+SESS_setVar('glWeather', $Session);
+
 $T->parse('output', 'index');
 $content .= $T->finish($T->get_var('output'));
 
@@ -111,25 +119,5 @@ $display .= $content;
 $display .= WEATHER_siteFooter();
 echo $display;
 exit;
-
-
-/**
-*   Strips slashes if magic_quotes_gpc is on.
-*
-*   @param  mixed   $var    Value or array of values to strip.
-*   @return mixed           Stripped value or array of values.
-*/
-function WEATHER_stripslashes($var)
-{
-	if (GVERSION < '1.3.0' && get_magic_quotes_gpc()) {
-		if (is_array($var)) {
-			return array_map('WEATHER_stripslashes', $var);
-		} else {
-			return stripslashes($var);
-		}
-	} else {
-		return $var;
-	}
-}
 
 ?>